requirements.txt
flask==3.0.3
pillow==10.4.0
numpy==2.1.1
pytest==8.3.3
pytest-cov==5.0.0